# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AvailabilityPromptPayload": {
        "additionalProperties": false,
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "posting_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Posting Ids",
            "type": "array"
          }
        },
        "required": [
          "as_of_date",
          "posting_ids"
        ],
        "title": "AvailabilityPromptPayload",
        "type": "object"
      },
...
      "FullResearchScope": {
        "properties": {
          "as_of_date": {
            "title": "As Of Date",
            "type": "string"
          },
          "type": {
            "const": "full_update",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "as_of_date"
        ],
        "title": "FullResearchScope",
        "type": "object"
      },
      "GapCriterionScopeRequest": {
        "additionalProperties": false,
        "properties": {
          "correlation_ref": {
            "title": "Correlation Ref",
            "type": "string"
          },
          "criterion_id": {
            "title": "Criterion Id",
            "type": "string"
          },
          "observation_type": {
            "title": "Observation Type",
            "type": "null"
          },
          "subject_type": {
            "enum": [
              "company",
              "opportunity",
              "posting"
            ],
            "title": "Subject Type",
            "type": "string"
          }
        },
        "required": [
          "subject_type",
          "correlation_ref",
          "criterion_id"
        ],
        "title": "GapCriterionScopeRequest",
        "type": "object"
      },
      "GapObservationScopeRequest": {
        "additionalProperties": false,
        "properties": {
          "correlation_ref": {
            "title": "Correlation Ref",
            "type": "string"
          },
          "criterion_id": {
            "title": "Criterion Id",
            "type": "null"
          },
          "observation_type": {
            "enum": [
              "technology_requirement",
              "task",
              "seniority",
              "experience_requirement",
              "work_model",
              "salary"
            ],
            "title": "Observation Type",
            "type": "string"
          },
          "subject_type": {
            "enum": [
              "company",
              "opportunity",
              "posting"
            ],
            "title": "Subject Type",
            "type": "string"
          }
        },
        "required": [
          "subject_type",
          "correlation_ref",
          "observation_type"
        ],
        "title": "GapObservationScopeRequest",
        "type": "object"
      },
      "GapRequestPayload": {
        "additionalProperties": false,
        "properties": {
          "criterion_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criterion Id"
          },
          "observation_type": {
            "anyOf": [
              {
                "enum": [
                  "technology_requirement",
                  "task",
                  "seniority",
                  "experience_requirement",
                  "work_model",
                  "salary"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observation Type"
          },
          "subject_id": {
            "minLength": 1,
            "title": "Subject Id",
            "type": "string"
          },
          "subject_type": {
            "enum": [
              "company",
              "opportunity",
              "posting"
            ],
            "title": "Subject Type",
            "type": "string"
          }
        },
        "required": [
          "subject_type",
          "subject_id"
        ],
        "title": "GapRequestPayload",
        "type": "object"
      },
      "GapResearchScope": {
        "properties": {
          "as_of_date": {
            "title": "As Of Date",
            "type": "string"
          },
          "requests": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/GapObservationScopeRequest"
                },
                {
                  "$ref": "#/components/schemas/GapCriterionScopeRequest"
                }
              ]
            },
            "title": "Requests",
            "type": "array"
          },
          "selected_correlation_refs": {
            "items": {
              "type": "string"
            },
            "title": "Selected Correlation Refs",
            "type": "array"
          },
          "type": {
            "const": "gap_filling",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "as_of_date",
          "selected_correlation_refs",
          "requests"
        ],
        "title": "GapResearchScope",
        "type": "object"
      },
      "GeneratedAvailabilityPromptResponse": {
        "properties": {
          "bundle_kind": {
            "const": "availability_check",
            "title": "Bundle Kind",
            "type": "string"
          },
          "bundle_version": {
            "const": "1.0",
            "title": "Bundle Version",
            "type": "string"
          },
          "prompt_context_ref": {
            "title": "Prompt Context Ref",
            "type": "string"
          },
          "prompt_run_id": {
            "title": "Prompt Run Id",
            "type": "string"
          },
          "prompt_text": {
            "title": "Prompt Text",
            "type": "string"
          },
          "prompt_type": {
            "const": "availability_check",
            "title": "Prompt Type",
            "type": "string"
          },
          "prompt_version": {
            "const": "1.0",
            "title": "Prompt Version",
            "type": "string"
          },
          "research_scope": {
            "additionalProperties": true,
            "title": "Research Scope",
            "type": "object"
          }
        },
        "required": [
          "prompt_run_id",
          "prompt_context_ref",
          "prompt_type",
          "prompt_version",
          "bundle_kind",
          "bundle_version",
          "research_scope",
          "prompt_text"
        ],
        "title": "GeneratedAvailabilityPromptResponse",
        "type": "object"
      },
      "GeneratedPromptResponse": {
        "properties": {
          "bundle_version": {
            "enum": [
              "1.0",
              "2.0"
            ],
            "title": "Bundle Version",
            "type": "string"
          },
          "criteria_count": {
            "title": "Criteria Count",
            "type": "integer"
          },
          "prompt_run_id": {
            "title": "Prompt Run Id",
            "type": "string"
          },
          "prompt_text": {
            "title": "Prompt Text",
            "type": "string"
          }
        },
        "required": [
          "prompt_run_id",
          "prompt_text",
          "bundle_version",
          "criteria_count"
        ],
        "title": "GeneratedPromptResponse",
        "type": "object"
      },
      "GeneratedUpdatePromptResponse": {
        "properties": {
          "bundle_version": {
            "const": "2.0",
            "title": "Bundle Version",
            "type": "string"
          },
          "criteria_count": {
            "title": "Criteria Count",
            "type": "integer"
          },
          "prompt_context_ref": {
            "title": "Prompt Context Ref",
            "type": "string"
          },
          "prompt_run_id": {
            "title": "Prompt Run Id",
            "type": "string"
          },
          "prompt_text": {
            "title": "Prompt Text",
            "type": "string"
          },
          "prompt_type": {
            "enum": [
              "full_update",
              "company_update",
              "opportunity_update",
              "gap_filling"
            ],
            "title": "Prompt Type",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "research_scope": {
            "discriminator": {
              "mapping": {
                "company_update": "#/components/schemas/SelectedResearchScope",
                "full_update": "#/components/schemas/FullResearchScope",
                "gap_filling": "#/components/schemas/GapResearchScope",
                "opportunity_update": "#/components/schemas/SelectedResearchScope"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/FullResearchScope"
              },
              {
                "$ref": "#/components/schemas/SelectedResearchScope"
              },
              {
                "$ref": "#/components/schemas/GapResearchScope"
              }
            ],
            "title": "Research Scope"
          }
        },
        "required": [
          "prompt_run_id",
          "prompt_context_ref",
          "prompt_type",
          "prompt_version",
          "bundle_version",
          "research_scope",
          "prompt_text",
          "criteria_count"
        ],
        "title": "GeneratedUpdatePromptResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InitialPromptPayload": {
        "additionalProperties": false,
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "title": "Constraints",
            "type": "array"
          },
          "search_profile": {
            "minLength": 1,
            "title": "Search Profile",
            "type": "string"
          }
        },
        "required": [
          "search_profile",
          "as_of_date"
        ],
        "title": "InitialPromptPayload",
        "type": "object"
      },
...
      "SelectedResearchScope": {
        "properties": {
          "as_of_date": {
            "title": "As Of Date",
            "type": "string"
          },
          "selected_correlation_refs": {
            "items": {
              "type": "string"
            },
            "title": "Selected Correlation Refs",
            "type": "array"
          },
          "type": {
            "enum": [
              "company_update",
              "opportunity_update"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "as_of_date",
          "selected_correlation_refs"
        ],
        "title": "SelectedResearchScope",
        "type": "object"
      },
...
      "UpdatePromptPayload": {
        "additionalProperties": false,
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "gap_requests": {
            "items": {
              "$ref": "#/components/schemas/GapRequestPayload"
            },
            "title": "Gap Requests",
            "type": "array"
          },
          "mode": {
            "enum": [
              "full_update",
              "company_update",
              "opportunity_update",
              "gap_filling"
            ],
            "title": "Mode",
            "type": "string"
          },
          "selected_ids": {
            "items": {
              "type": "string"
            },
            "title": "Selected Ids",
            "type": "array"
          }
        },
        "required": [
          "mode",
          "as_of_date"
        ],
        "title": "UpdatePromptPayload",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Vocation",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/prompts/availability-check": {
      "post": {
        "operationId": "generate_availability_prompt_api_prompts_availability_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityPromptPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedAvailabilityPromptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Availability Prompt",
        "tags": [
          "research prompts"
        ]
      }
    },
    "/api/prompts/initial": {
      "post": {
        "operationId": "generate_initial_prompt_api_prompts_initial_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitialPromptPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedPromptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Initial Prompt",
        "tags": [
          "research prompts"
        ]
      }
    },
    "/api/prompts/update": {
      "post": {
        "operationId": "generate_update_prompt_api_prompts_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromptPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedUpdatePromptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Update Prompt",
        "tags": [
          "research prompts"
        ]
      }
    },
...
  }
}